+++ tests/flows/test_fpgaflow.py
@@ -1,10 +1,12 @@
 import os
+import json
 import pytest
 import siliconcompiler
 from siliconcompiler.scheduler import _setup_node
 from siliconcompiler.targets import fpgaflow_demo
 from siliconcompiler.tools.vpr import route, place
 from siliconcompiler.flowgraph import _get_flowgraph_execution_order
+from siliconcompiler.utils import register_sc_data_source
 
 
 @pytest.mark.eda
@@ -490,6 +492,93 @@ def test_vpr_max_router_iterations(scroot,
     chip.set('arg', 'step', 'route')
     chip.set('arg', 'index', '0')
     assert f'--max_router_iterations {test_value}' in route.runtime_options(chip)
+
+
+@pytest.mark.eda
+@pytest.mark.quick
+@pytest.mark.parametrize("top_module, expected_macro",
+                         [('macc_pipe', 'efpga_mult'),
+                          ('adder_extract', 'efpga_adder')])
+def test_fpga_syn_extract(top_module,
+                          expected_macro,
+                          datadir,
+                          examples_root):
+
+    # Build FPGA
+    arch_name = 'example_arch_test_fpgasyn'
+
+    fpga = siliconcompiler.FPGA(arch_name, package='siliconcompiler_data')
+    register_sc_data_source(fpga)
+
+    # Set the absolute minimum number of things needed to run
+    # synthesis tests (add other properties as needed when writing new tests)
+
+    fpga.set('fpga', arch_name, 'lutsize', 4)
+    fpga.add('fpga', arch_name, 'var', 'feature_set', 'async_reset')
+    fpga.add('fpga', arch_name, 'var', 'feature_set', 'async_set')
+    fpga.add('fpga', arch_name, 'var', 'feature_set', 'enable')
+
+    techlib_root = os.path.join(datadir, 'test_fpgasyn')
+
+    mae_library = os.path.join(techlib_root, 'tech_mae.v')
+    fpga.add('fpga', arch_name, 'file', 'yosys_extractlib', mae_library)
+    fpga.add('fpga', arch_name, 'file', 'yosys_macrolib', mae_library)
+
+    # Setup chip
+    chip = siliconcompiler.Chip(top_module)
+
+    chip.set('fpga', 'partname', arch_name)
+
+    chip.set('option', 'to', 'syn')
+
+    flow_root = os.path.join(examples_root, 'fpga_flow')
+
+    # 1. Defining the project
+    # 2. Define source files
+    v_src = os.path.join(flow_root, 'designs', top_module, f'{top_module}.v')
+    chip.input(v_src)
+
+    # 3. Load target
+    chip.use(fpga)
+    chip.use(fpgaflow_demo)
+
+    assert chip.check_filepaths()
+
+    chip.run()
+
+    report_file = f'build/{top_module}/job0/syn/0/reports/stat.json'
+    assert os.path.isfile(report_file)
+
+    with open(report_file, "r") as report_data:
+        stats = json.loads(report_data.read())
+
+        # In Yosys 0.44 at least, the stats are presented in two places.  One
+        # is under a two-level keypath 'modules', '\\<top_module_name>';
+        # and the other is at a top level key called 'design'
+        # choose 'design' to key on so we're digging through one less
+        # level of JSON dictionary hierarchy
+        assert 'design' in stats, 'stats for top level "design" not found'
+
+        design_stats = stats['design']
+
+        # Breakdown of cell counts by type live within a key called
+        # 'num_cells_by_type'
+
+        assert 'num_cells_by_type' in design_stats, 'num_cells_by_type dictionary not found'
+
+        cells_by_type = design_stats['num_cells_by_type']
+
+        # Set up this test so that we are looking for exactly one
+        # instance of a particular extracted macro
+
+        assert expected_macro in cells_by_type, \
+            f'Expected macro {expected_macro} not found in cell types report'
+
+        expected_macro_count = cells_by_type[expected_macro]
+
+        assert expected_macro_count == 1, \
+            f'Expected one instance of {expected_macro},' \
+            ' got {expected_macro_count} instances'
 
 
 if __name__ == "__main__":
